sqrt(weight) dtype not same as weight

+++ src/aspire/sinogram/sinogram.py
@@ -37,7 +37,7 @@ class Sinogram:
                 f"Invalid data shape: {data.shape}. Expected shape: (..., angles, radial_points), where '...' is the stack number."
             )
 
-        self._data = data.astype(self.dtype, copy=None)
+        self._data = data.astype(self.dtype, copy=False)
         self.ndim = self._data.ndim
         self.shape = self._data.shape
         self.stack_shape = self._data.shape[:-2]
